feat: clean up extra tests
